Lab5 trying to save users in database.
Implemented the database connection code, resolving all exceptions, but
I think I will circle back to the idea of extending the
"InMemoryRepository", and then work on the memory data, not the
database.

Because of the incapability to inherit from multiple classes, I thought
of making a static function that establishes connection to the database,
But now when I think about it, with the new idea of extending
"InMemoryRepository", I only need to establish the connection 2 times,
when loading and when saving.

I am very confused.