Lab 6 for MAP. Saving in database in progress
I came to a conclusion: implement a connectToDb function that returns a
connection to the db with the given credentials(properties of the
abstractDataRepository class). And each method, like save, findOne,
findAll, delete, update, will call the connection method in order to
apply changes or interogate the database.

I find it easier to interact with the database for every insert or
information needed, and I also chose to make the application this way
because it would be so unnecessary to load the whole database in memory,
in order to find/save/delete.

Problems: the find function doesn't resolve all exceptions and closings.

Not implemented:
-findAll (SELECT * FROM USERS)
-save (
	SET IDENTITY_INSERT USERS ON;
	INSERT INTO USERS (id, first_name, last_name) VALUES
		(1, "Alex", "Zdroba")
	SET IDENTITY_INSERT USERS OFF;)
-delete(DELETE FROM USERS WHERE USERS.id = ?)
-I am contemplating the idea of adding the update function as I don't
think I even implemented it in the UI